Restructure the clo source files into the clp namespace.
